Remove input variable coercion since it is done again later
When an incoming query is processed input coercion and  validation is
run on all inputs.

Since variables are coerced and validated separately, when the input
is processed the types have already been coerced and validation fails
if they are object types.

This removes variable coercion and follows the same process for all inputs.