Updated fix for nested objects as variables
- Removed deserialization from variable coercion and put it in the
argument coercion instead.  This keeps validation from being done
later in coercion of arguments on the coerced type and instead on the
json like it is expecting.  This would happen when passing objects
that are nested in another object as variables

- updated tests to use custom serializers when submitting a nested
object as a variable

- updated GraphQLListTypes to deserialize List<dynamic> instead of
requiring it to be the desired type.  The type is still cast to the
desired type in the deserialization.  This allows raw json to be
passed to a primitive list.